Added HeadNavComponent to use on all pages
