+++ app/src/components/head-nav/head-nav.module.ts
@@ -1,0 +1,16 @@
+import { NgModule } from '@angular/core';
+import { IonicModule } from 'ionic-angular';
+import { HeadNavComponent } from './head-nav';
+
+@NgModule({
+  declarations: [
+    HeadNavComponent,
+  ],
+  imports: [
+    IonicModule,
+  ],
+  exports: [
+    HeadNavComponent
+  ]
+})
+export class HeadNavComponentModule {}
