+++ app/src/components/head-nav/head-nav.scss
@@ -1,0 +1,3 @@
+head-nav {
+
+}
