+++ app/src/pages/address/address.html
@@ -1,17 +1,6 @@
-<!--
-  Generated template for the AddressPage page.
-
-  See http://ionicframework.com/docs/components/#navigation for more info on
-  Ionic pages and navigation.
--->
 <ion-header>
-
-  <ion-navbar>
-    <ion-title>Address</ion-title>
-  </ion-navbar>
-
+  <head-nav [title]="'Address'"></head-nav>
 </ion-header>
-
 
 <ion-content padding>
   <ion-grid>
