+++ app/src/pages/block-detail/block-detail.html
@@ -1,17 +1,6 @@
-<!--
-  Generated template for the BlockDetailPage page.
-
-  See http://ionicframework.com/docs/components/#navigation for more info on
-  Ionic pages and navigation.
--->
 <ion-header>
-
-  <ion-navbar color="primary">
-    <ion-title>Block Detail</ion-title>
-  </ion-navbar>
-
+  <head-nav [title]="'Block'"></head-nav>
 </ion-header>
-
 
 <ion-content padding>
   <h1>Block # {{ block.height }}</h1>
