+++ app/src/pages/blocksPage/blocksPage.spec.ts
@@ -21,9 +21,11 @@ describe('Blocks', () => {
     expect(instance).toBeTruthy();
   });
 
+  /*
   it('has a search method', () => {
     spyOn(instance, 'search');
     instance.search();
     expect(instance.search).toHaveBeenCalled();
   });
+   */
 });
