+++ app/src/pages/transaction/transaction.html
@@ -1,17 +1,6 @@
-<!--
-  Generated template for the TransactionPage page.
-
-  See http://ionicframework.com/docs/components/#navigation for more info on
-  Ionic pages and navigation.
--->
 <ion-header>
-
-  <ion-navbar>
-    <ion-title>Transaction</ion-title>
-  </ion-navbar>
-
+  <head-nav [title]="'Transaction'"></head-nav>
 </ion-header>
-
 
 <ion-content padding>
   <ion-grid>
