+++ app/src/pages/transaction/transaction.module.ts
@@ -2,6 +2,7 @@ import { NgModule } from '@angular/core';
 import { IonicPageModule } from 'ionic-angular';
 import { TransactionPage } from './transaction';
 import { TransactionComponentModule } from '../../components/transaction/transaction.module';
+import { HeadNavComponentModule } from '../../components/head-nav/head-nav.module';
 
 @NgModule({
   declarations: [
@@ -9,7 +10,8 @@ import { TransactionComponentModule } from '../../components/transaction/transac
   ],
   imports: [
     IonicPageModule.forChild(TransactionPage),
-    TransactionComponentModule
+    TransactionComponentModule,
+    HeadNavComponentModule
   ],
   exports: [
     TransactionPage
